KCL refactor: Move execution into its own file
types.rs is really, really big. I've moved executing AST nodes into
their own file.
